refactor(AppSearch): delete AppSearch code related to venues
