+++ src/libs/search/filters/buildBoosts.ts
@@ -3,7 +3,7 @@ import { Boosts, ProximityBoost } from '@elastic/app-search-javascript'
 import { LocationFilter } from 'features/search/types'
 import { GeoCoordinates } from 'libs/geolocation'
 
-import { AppSearchFields, AppSearchVenuesFields } from './constants'
+import { AppSearchFields } from './constants'
 
 const getProximityBoost = (userLocation: GeoCoordinates): ProximityBoost => ({
   type: 'proximity',
@@ -21,12 +21,4 @@ export const buildBoosts = (
 ): Boosts<AppSearchFields> | undefined => {
   if (!userLocation || !isSearchFilteredAroundUserPosition(locationFilter)) return
   return { [AppSearchFields.venue_position]: getProximityBoost(userLocation) }
-}
-
-export const buildBoostsVenues = (
-  userLocation: GeoCoordinates | null,
-  locationFilter: LocationFilter
-): Boosts<AppSearchVenuesFields> | undefined => {
-  if (!userLocation || !isSearchFilteredAroundUserPosition(locationFilter)) return
-  return { [AppSearchVenuesFields.position]: getProximityBoost(userLocation) }
 }
