+++ src/libs/search/filters/index.ts
@@ -1,3 +1,2 @@
 export { AppSearchFields, RESULT_FIELDS } from './constants'
 export { underageFilter, buildQueryOptions } from './buildQueryOptions'
-export { buildVenuesQueryOptions } from './buildVenuesQueryOptions'
